Move `std::io::BufRead` to `alloc::io`
