+++ library/alloc/src/io/mod.rs
@@ -1,5 +1,6 @@
 //! Traits, helpers, and type definitions for core I/O functionality.
 
+mod buf_read;
 mod cursor;
 mod error;
 mod impls;
@@ -27,6 +28,12 @@ pub use core::io::{
     stream_len_default, take,
 };
 
+#[unstable(feature = "alloc_io", issue = "154046")]
+pub use self::{
+    buf_read::BufRead,
+    read::{Read, read_to_string},
+    util::{Bytes, Lines, Split},
+};
 #[doc(hidden)]
 #[unstable(feature = "core_io_internals", reason = "exposed only for libstd", issue = "none")]
 pub use self::{
@@ -34,10 +41,5 @@ pub use self::{
         DEFAULT_BUF_SIZE, append_to_string, default_read_buf, default_read_buf_exact,
         default_read_exact, default_read_to_end, default_read_to_string, default_read_vectored,
     },
-    util::{SpecReadByte, bytes, uninlined_slow_read_byte},
-};
-#[unstable(feature = "alloc_io", issue = "154046")]
-pub use self::{
-    read::{Read, read_to_string},
-    util::Bytes,
+    util::{SpecReadByte, bytes, lines, split, uninlined_slow_read_byte},
 };
